fix: Disable local login in admin page #384
